hookify getting local samples logic
