+++ src/components/App.tsx
@@ -3,14 +3,12 @@ import Providers from './Providers';
 import Sampler from './Sampler';
 import SamplesLoader from './SamplesLoader';
 import MidiConnector from './MidiConnector';
-import SamplesFetcher from './SamplesFetcher';
 
 const App = () => (
   <>
     <h1 className="header">boomTap</h1>
     <Providers>
       <>
-        <SamplesFetcher />
         <SamplesLoader />
         <Sampler />
         <MidiConnector />
